플래시카드 플레이 데이터, 검색 / 레코드 추가, 변경, 통계 기능;

+++ src/main/java/btongtong/btongtalkback/repository/CategoryRepository.java
@@ -24,6 +24,6 @@ public interface CategoryRepository extends JpaRepository<Category, Long> {
             "LEFT JOIN c.flashcards f " +
             "WHERE c.parent.id = :parentId " +
             "GROUP BY c.id, c.name")
-    List<CategoryDto> findCategories(@Param("parentId") int parentId);
+    List<CategoryDto> findCategories(@Param("parentId") Long parentId);
 
 }
